+++ src/i18n.rs
@@ -12,8 +12,8 @@ fluent_templates::static_loader! {
 pub const SUPPORTED_LANGUAGES: &[LanguageIdentifier] = &[
     langid!("en-us"),
     langid!("ru-ru"),
-    langid!("zh-cn"),
-    langid!("pt-br")
+    langid!("pt-br"),
+    langid!("zh-cn")
 ];
 
 static mut LANG: LanguageIdentifier = langid!("en-us");